커밋

+++ Client/Assets/__Script/Pawn/PawnBlackboardHelper.cs
@@ -5,13 +5,13 @@ namespace Game
 {
     public static class PawnBlackboardHelper
     {
-        public static int ConsumeActionPoint(this PawnBlackboard.Stat stat, int delta)
+        public static float ConsumeActionPoint(this PawnBlackboard.Stat stat, int delta)
         {   
             stat.actionPoint.Value = Math.Max(0, stat.actionPoint.Value - delta);
             return stat.actionPoint.Value;
         }
 
-        public static int RecoverActionPoint(this PawnBlackboard.Stat stat, int delta)
+        public static float RecoverActionPoint(this PawnBlackboard.Stat stat, int delta)
         {   
             stat.actionPoint.Value = Math.Min(stat.maxActionPoint.Value, stat.actionPoint.Value + delta);
             return stat.actionPoint.Value;
